requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.0
facebook-sdk==3.0.0
Pillow==10.1.0
python-dateutil==2.8.2
PyPDF2==3.0.1
python-docx==0.8.11
psycopg2-binary==2.9.9
gunicorn==21.2.0
cryptography==41.0.7
PyJWT==2.10.1
anthropic==0.75.0
